add os_helper submodule in test.support
